restructured js to make more sense semantically and adjusted variables
Reorganized and cleaned up js file so it fits with semantic standards. App now functions how it should on a basic level. Added all the values as a starting point for a password generator that utilizes all the possible combinations when mixing different character types: uppercase, lowercase, special, numeric. All that's left is to create prompts asking user about each element that they desire in the password and implementing logic that also applies accordingly with user criteria.